metric/export : Reorganize example/metric folders.
