--- example/metric/collector/sample-collector-config.yaml
@@ -1,40 +1,0 @@
-# This is a sample config file which can be used with an opentelemetry collector
-# This sample is configured to recieve OTLP metrics over HTTP on localhost:4318
-# and export the recieved metrics to a Google Cloud project.
-receivers:
-  otlp:
-    protocols:      
-      http:
-        endpoint: "localhost:4318"
-
-processors:
-  memory_limiter:
-    check_interval: 1s
-    limit_percentage: 65
-    spike_limit_percentage: 20
-  batch:
-
-exporters:
-  googlecloud:    
-    retry_on_failure:
-      enabled: false  
-
-extensions:
-  health_check:
-  pprof:
-  zpages:
-
-service:
-  telemetry:
-    metrics:
-      address: ":8888"
-  extensions: [health_check, pprof, zpages]
-  pipelines:
-    metrics:
-      receivers: [otlp]
-      processors: [memory_limiter, batch]
-      exporters: [googlecloud]  
-    logs:
-      receivers: [otlp]
-      processors: [memory_limiter, batch]
-      exporters: [googlecloud]
