--- example/metric/create_dashboard.sh
@@ -1,20 +1,0 @@
-#!/usr/bin/env bash
-
-# Copyright 2020 Google LLC
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-# This dashboard is intended to show the metrics emit from the example
-# in the same directory.
-# https://cloud.google.com/monitoring/dashboards/api-dashboard
-gcloud monitoring dashboards create --config-from-file=dashboard.json
