--- example/metric/dashboard.json
@@ -1,67 +1,0 @@
-{
-  "displayName": "OpenTelemetry exporter example/metric",
-  "gridLayout": {
-    "columns": "1",
-    "widgets": [
-      {
-        "title": "custom/opentelemetry/counter-a",
-        "xyChart": {
-          "chartOptions": {
-            "mode": "COLOR"
-          },
-          "dataSets": [
-            {
-              "minAlignmentPeriod": "60s",
-              "plotType": "LINE",
-              "timeSeriesQuery": {
-                "timeSeriesFilter": {
-                  "aggregation": {
-                    "crossSeriesReducer": "REDUCE_SUM",
-                    "perSeriesAligner": "ALIGN_DELTA"
-                  },
-                  "filter": "metric.type=\"workload.googleapis.com/counter-a\"",
-                  "secondaryAggregation": {}
-                },
-                "unitOverride": "1"
-              }
-            }
-          ],
-          "timeshiftDuration": "0s",
-          "yAxis": {
-            "label": "total count",
-            "scale": "LINEAR"
-          }
-        }
-      },
-      {
-        "title": "custom/opentelemetry/observer-a",
-        "xyChart": {
-          "chartOptions": {
-            "mode": "COLOR"
-          },
-          "dataSets": [
-            {
-              "minAlignmentPeriod": "60s",
-              "plotType": "LINE",
-              "timeSeriesQuery": {
-                "timeSeriesFilter": {
-                  "aggregation": {
-                    "perSeriesAligner": "ALIGN_MEAN"
-                  },
-                  "filter": "metric.type=\"workload.googleapis.com/observer-a\"",
-                  "secondaryAggregation": {}
-                },
-                "unitOverride": "1"
-              }
-            }
-          ],
-          "timeshiftDuration": "0s",
-          "yAxis": {
-            "label": "guage value",
-            "scale": "LINEAR"
-          }
-        }
-      }
-    ]
-  }
-}
